requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
plotly>=5.21.0
joblib>=1.3.2
fpdf>=1.7.2
shap>=0.45.0
Pillow>=10.3.0
